Solution to Check LSB of a Number

+++ Check LSB of a Number@@67882a731b4392d8fc1577e8/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 #include<ctype.h>
 int main(){
-    long int a;
+    unsigned int a;
     scanf("%d",a);
     if(a==1){
         printf("Set");
